Support context menu options
Add in cursor position to reveal definition
